+++ CHANGELOG.md
@@ -15,7 +15,7 @@ The rules for this file:
     * YYYY-MM-DD date format (following ISO 8601)
   * accompany each entry with github issue/PR number (Issue #xyz)
 -->
-## [1.1.4] - unpublished
+## [1.2.0] - unpublished
 
 ### Authors
 * @orbeckst
@@ -28,7 +28,7 @@ The rules for this file:
   whole proteins.
 
 ### Added
-* Added command-line interface for basicrta workflow
+* Added command-line interface for basicrta workflow (Issue #20)
 
 ## [1.1.3] - 2025-09-11